Fix for http://github.com/AsyncHttpClient/async-http-client/issues/issue/44 ("Race condition when the remote server close the connection and ConnectionsPool is used")
When the remote server close the connection, there is an interval where Netty still didn't detect the connection has been closed and the client try to re-use that connection from the connection pool. In that case the connection is no longer valid and we need to reconnect using a new one.