Handles params & queryParams only if maps are not empty
Params were set in the request even if the map of params was empty, this could lead to
requests with no body even if we set one.
This commit fixes this behaviour for Netty provider.
The same has been done for queryParams (in RequestBuilderBase) to prevent '?' to be added even if not wanted