Ensured relative URIs are used to connect to proxied WebSockets
This needed to be done both for the Grizzly and the Netty provider.

Since this is essentially the same configuration parameter as the
useRelativeURIsWithSSLProxies, I reused that parameter, but I renamed it
to useRelativeURIsWithConnectProxies to reflect its true meaning since
it's no longer used just for SSL, and I deprecated the old parameter.