fix to enable client-specified nonce count (NC)
Current implementation seems to contain a minor oversight. The static final (default) field `NC` is used when calculating the response, instead of the member variable `nc`, so the latter is ignored here (and, dangerously, is not ignored when composing the Digest header in `AuthenticatorUtils.computeDigestAuthentication()`

Existing functionality is not affected since we maintain the original default (renamed the constant field for clarity).