AsyncHttpClient `name` option
Fixes #952

Name option is useful for two purposes:

## Thread name

Typical application with several netty and/or async-http-client
instances may have hundreds of threads like "New I/O worker 120",
which makes it hard to understand which thread belongs to which
part of the application.

Specified name is used in AsyncHttpClient to name threads created
by AsyncHttpClient and netty.

If name is not specified, `AsyncHttpClient` string is used.

## Debugging

`name` is stored in `AsyncHttpClientConfig`, and available in
debugger debugging when application has multiple client instances.
It can be useful to distinguish between instance, for example, when
breakpoint is set inside of AsyncHttpClient.