Implement correct handling of recursive DNS, close #1329
Motivation:

DnsNameResolver does not handle recursive DNS and so fails if you query
a DNS server (for example a ROOT dns server) which provides the correct
redirect for a domain.

Modification:

Add support for redirects (a.k.a. handling of AUTHORITY section').

Result:

Its now possible to use a DNS server that redirects.