Drop WebSocketUpgradeHandler:: touchSuccess
Motivation

This hack was a very old workaround for #317 where the server would
send the first frame along with the handshake.
This issue has actually been fixed for quite some time, we just had to
replace the `UpgradeCallback` with the future before triggering the
read when upgrading the pipeline.

We no longer need this protection as there’s no way the
`UpgradeCallback` can be called twice.

Modification:

Remove dead code.

Result:

Clean code