Don't upgrade pipeline for WebSocket prior to installing WebSocket on Handler, close #1364
Motivation:

Installing the WebSocket handler can trigger a read, so if there's a
pending frame that was send along the HTTP upgrade response, we could
crash because the handler is not properly configured yet.

This is a regression caused by #1348.

Modifications:

Buffer frames until WebSocket is fully open.

Result:

No more crash when frames are sent along the HTTP upgrade response.