Fix broken Head302Test (#1421)
* Update Head302Test to show it is actually failing

This test had the following defects which caused it to falsely pass:

* Jetty requires a handler to explicitly indicate it has handled the
request by calling Request.setHandled(true).  Otherwise the server will
return 404.
* The test did not configure the client to follow redirects.
* The test itself wasn't asserting anything useful beyond that the
request did not time out.  To ensure the redirect was followed, it needs
to assert the expected response status code from the redirected location
is received.

* Fix broken Head302Test

The test now verifies that a HEAD redirected via 302 is switched to GET
per [1].

[1] https://github.com/AsyncHttpClient/async-http-client/issues/989