Expect body chunks when aborting during status handling, close #1430
Motivation:

When aborting during response status handling, we currently directly
offer the channel to the pool. This is wrong as we’ll at least receive
the LastHttpContent. Currently, this event is received by the next
request being executed with this channel, causing a null response.

Modification:

Always expect chunks when aborting during status or headers handling,
causing channel to be drained before being offered to the pool.

Note that AHC 2.1 doesn’t suffer from this issue as aborting now always
close the channel.

Result:

No more null response when aborting on status or headers handling