Fix form url encoding when charset is not UTF-8, close #1444
Motivation:

form urlencoding doesn’t properly honor charset. It uses it for
converting the bytes while those are supposed to be already in the
US-ASCII range. It should be using it the first encode into bytes,
which should be then escaped.

Modifications:

Use current optimized code for UTF-8 and fall back to URLEncoder for
other charsets.

Results:

Proper encoding when charset is different from UTF-8, eg GBK