Perform WriteListener#abortOnThrowable whatever the channel state, close #1488
Motivation

For some reason lost in time, abortOnThrowable is only performed if the
channel is not a new one.
I have no idea why and it swallow exceptions, so we get timeouts
instead of actual exception.

Modification:

Drop this test and always perform abortOnThrowable, whatever the
channel state.

Result:

Proper exception reported