Force Content-Length to 0 when there's no body but method typically uses one, close #1521
Motivation:

From RFC7230:

A user agent SHOULD send a Content-Length in a request message when
no Transfer-Encoding is sent and the request method defines a meaning
for an enclosed payload body. For example, a Content-Length header
field is normally sent in a POST request even when the value is 0
(indicating an empty payload body). A user agent SHOULD NOT send a
Content-Length header field when the request message does not contain
a payload body and the method semantics do not anticipate such a
body.

Some servers, such as IIS, don’t support POST requests without
Content-Length nor transfer encoding.

Modification:

Force a 0 Content-Length for PUT, POST and PATCH requests without a
body.

Result:

No more 411 Length Required