Don't set peer domain and port in SSLEngine when HttpsEndpointIdentificationAlgorithm is disabled, close #1535
Motivation:

Some broken TLS implementations crash on SSL handshakes that have host
and port. Browers handle this by retrying without those parameters, but
that’s definitively not something we’ll implement.

Disabling HttpsEndpointIdentificationAlgorithm should suffice to be
able to connect to such broken server.

Modifications:

When HttpsEndpointIdentificationAlgorithm is disabled (that disables
SNI and hostname verification), we want to not set peer domain and port
on the SSLEngine so those are not sent in the SSL handshake and make
broken servers crash.

Result:

It’s now possible to connect to broken servers that don’t support
hostname and port in SSL handshake.