Trigger retry in WriteListener on SSLException, close #1645
Motivation:

There's a possibility that we have a write failure because we're trying to write on a pooled channel and the remote peer has closed the SSLEngine but we haven't process the connection close yet.

In this case, we don't trigger retry, and don't forcefully close the channel.

Modifications:

* Trigger retry on write failure on SSLException.
* Always close channel on write failure

Result:

Retry triggered on SSLEngine crash.
No more pooled connection in stalled state.