Allow disabling automatic decompression.
Adds new config option "isEnableAutomaticDecompression()" to config and builders.

If set to true (default), behavior is as prior: Automatic decompression of compressed content by Netty HTTP client.

If set to false, the content will not get decompressed automatically, but is passed to AsyncHandler as retrieved from the server.

This is independent of config option isCompressionEnforced(). Added comments, to make this clear.

Committed again after merging upstream and resolving conflicts.