fix: send CONNECT first when recovering a HTTPS request (#2077)
# Issue description

AHC has retry mechanism enabled with up to 5 attempts by default. But
the initial CONNECT is omitted when recovering the HTTPS requests with
IO exceptions. This MR fixes this issue and guarantees the proper
workflow in retries.

It's related to #2071 and fixes a different failing case.

# How the issue is fixed

* For any new connections, make sure there is an initial CONNECT for
WebSocket/HTTPS request.
* For the condition check that a CONNECT has been sent, make sure the
connection the current future attaches is reusable/active.

# Unit test

IOException has various reasons but in the unit test, we emulate it by
closing the connection after receiving the CONNECT request. The internal
recovery process will retry another 4 times, and through an IOException
eventually.

Signed-off-by: Jason Joo <hblzxsj@gmail.com>