Future-proof HTTPS endpoint identification (#2104)
Netty 4.2 changes the default for hostname verification for TLS clients,
so that it is now enabled by default.

As a result, clients that rely on the default being _off_ will find
themselves unable to disable it.

Instead, clients should explicitly configure their desired endpoint
identification algorithm in all cases.

Since Netty 4.1.112 we also have a convenient method on the
`SslContextBuilder` for doing this, so we don't need multiple
round-trips through `SSLParameters`.

This PR changes the `DefaultSslEngineFactory` to make use of this
method, so it always configures the endpoint identification algorithm to
match the desired setting of
`AsyncHttpClientConfig..isDisableHttpsEndpointIdentificationAlgorithm()`.